Reorganize transforms
