+++ fastai_sparse/version.py
@@ -1,2 +1,2 @@
 __all__ = ['__version__']
-__version__ = '0.0.1.dev0'
+__version__ = '0.0.2.dev0'
